+++ docs/apireference.rst
@@ -14,6 +14,7 @@ API Reference
       opentelemetry.exporter
       opentelemetry.exporter.google
       opentelemetry.tools
+      opentelemetry.tools.google
 
 
 
@@ -25,3 +26,4 @@ API Reference
    _autosummary/opentelemetry.exporter
    _autosummary/opentelemetry.exporter.google
    _autosummary/opentelemetry.tools
+   _autosummary/opentelemetry.tools.google
