+++ docs/examples/tools/cloud_trace_propagator/README.rst
@@ -19,6 +19,7 @@ To use this feature you first need to:
     pip install opentelemetry-api
     pip install opentelemetry-sdk
     pip install opentelemetry-exporter-google-cloud
+    pip install opentelemetry-tools-google-cloud
 
 * Create a server that uses the Cloud Trace header
