+++ docs/examples/tools/cloud_trace_propagator/client.py
@@ -17,12 +17,12 @@ import opentelemetry.ext.requests
 import requests
 from opentelemetry import trace
 from opentelemetry.exporter.cloud_trace import CloudTraceSpanExporter
-from opentelemetry.exporter.cloud_trace.cloud_trace_propagator import (
-    CloudTraceFormatPropagator,
-)
 from opentelemetry.propagators import set_global_httptextformat
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import SimpleExportSpanProcessor
+from opentelemetry.tools.cloud_trace_propagator import (
+    CloudTraceFormatPropagator,
+)
 
 # Instrumenting requests
 opentelemetry.ext.requests.RequestsInstrumentor().instrument()
