+++ docs/index.rst
@@ -9,6 +9,9 @@ Welcome to Google Cloud OpenTelemetry's documentation!
 .. image:: https://badge.fury.io/py/opentelemetry-exporter-google-cloud.svg
     :target: https://badge.fury.io/py/opentelemetry-exporter-google-cloud
 
+.. image:: https://badge.fury.io/py/opentelemetry-tools-google-cloud.svg
+    :target: https://badge.fury.io/py/opentelemetry-tools-google-cloud
+
 This documentation describes OpenTelemetry Python exporters, propagators, and
 resource detectors for Google Cloud Platform. Development for these packages
 takes place on `Github
@@ -17,9 +20,18 @@ takes place on `Github
 Installation
 ------------
 
+To install trace and metrics exporters:
+
 .. code-block:: bash
 
     pip install opentelemetry-exporter-google-cloud
+
+To install the additional tools package with resource detection and GCP trace
+context propagation:
+
+.. code-block:: bash
+
+    pip install opentelemetry-tools-google-cloud
 
 
 .. toctree::
