+++ opentelemetry-tools-google-cloud/CHANGELOG.md
@@ -1,0 +1,6 @@
+# Changelog
+
+## Unreleased
+
+- Add support for resources
+  ([#853](https://github.com/open-telemetry/opentelemetry-python/pull/853))
