+++ opentelemetry-tools-google-cloud/README.rst
@@ -1,0 +1,31 @@
+OpenTelemetry Google Cloud Integration
+======================================
+
+.. image:: https://badge.fury.io/py/opentelemetry-tools-google-cloud.svg
+    :target: https://badge.fury.io/py/opentelemetry-tools-google-cloud
+
+.. image:: https://readthedocs.org/projects/google-cloud-opentelemetry/badge/?version=latest
+    :target: https://google-cloud-opentelemetry.readthedocs.io/en/latest/?badge=latest
+    :alt: Documentation Status
+
+This library provides support for:
+
+- Detecting GCP resources like GCE, GKE, etc.
+- Propagating trace context in the legacy Google Cloud format
+
+Installation
+------------
+
+.. code:: bash
+
+    pip install opentelemetry-tools-google-cloud
+
+..
+    TODO: Add usage info here
+
+
+References
+----------
+
+* `Cloud Trace <https://cloud.google.com/trace/>`_
+* `OpenTelemetry Project <https://opentelemetry.io/>`_
