+++ opentelemetry-tools-google-cloud/setup.cfg
@@ -1,11 +1,11 @@
 [metadata]
-name = opentelemetry-exporter-google-cloud
-description = Google Cloud integrations for OpenTelemetry
+name = opentelemetry-tools-google-cloud
+description = Google Cloud propagator and resource detector for OpenTelemetry
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 author = Google
 author_email = opentelemetry-pypi@google.com
-url = https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/opentelemetry-exporter-google-cloud
+url = https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/opentelemetry-tools-google-cloud
 platforms = any
 license = Apache-2.0
 classifiers =
@@ -26,10 +26,10 @@ package_dir=
     =src
 packages=find_namespace:
 install_requires =
-    google-cloud-monitoring
-    google-cloud-trace
+    google-auth
     opentelemetry-api
     opentelemetry-sdk
+    requests
 
 [options.packages.find]
 where = src
