+++ opentelemetry-tools-google-cloud/src/opentelemetry/tools/cloud_trace_propagator.py
