+++ opentelemetry-tools-google-cloud/src/opentelemetry/tools/resource_detector.py
