+++ opentelemetry-tools-google-cloud/tests/test_gcp_resource_detector.py
