+++ tox.ini
@@ -3,7 +3,7 @@ skipsdist = True
 skip_missing_interpreters = True
 envlist =
   ; Add the `ci` factor to any env that should be running during CI.
-  py3{4,5,6,7,8}-ci-test-exporter-google-cloud
+  py3{4,5,6,7,8}-ci-test-{exporter,tools}-google-cloud
 
   ; These are development commands and use the same virtualenv.
   dev
@@ -30,6 +30,7 @@ dev_deps =
   isort
   pylint
   -e {toxinidir}/opentelemetry-exporter-google-cloud
+  -e {toxinidir}/opentelemetry-tools-google-cloud
 
 lint_commands =
   black . --diff --check
@@ -49,7 +50,8 @@ deps =
   test: pytest
 
 changedir =
-  test: opentelemetry-exporter-google-cloud
+  test-exporter: opentelemetry-exporter-google-cloud
+  test-tools: opentelemetry-tools-google-cloud
 
 commands_pre =
   test: pip install .
